feat: Move and change rootMargin
Move all the sections to a folder for easier lookup.

Change props for the sections to make it easier to understand what each variable is.

Add check for if page height is not set because it takes at least one render for the variable to be set.

Change rootMargin for the IntersectionObserver to be 1px wide strip in the center of the screen. This prevents the observer from being triggered several times and solves conflicts between two sections triggering the observer at the same time.

Add a useEffect for updating the active section to prevent infinte render loops.